+++ scripts/hooks.js
@@ -30,6 +30,7 @@ import  SR5MeasuredTemplateConfig  from "./interface/measuredTemplateConfig.js";
 import {SR5CompendiumInfo} from "./interface/compendium.js";
 import * as macros from "./interface/macros.js";
 import Migration from "./migration.js";
+import { SR5_ActorHelper } from "./entities/actors/entityActor-helpers.js";
 
 export const registerHooks = function () {
 	Hooks.once("init", async function () {
@@ -217,6 +218,7 @@ export const registerHooks = function () {
 	});
 
 	Hooks.on("createToken", async function(tokenDocument) {
+		if (!game.user.isGM) return;
 		let tokenData = duplicate(tokenDocument);
 		if (tokenDocument.texture.src.includes("systems/sr5/img/actors/")) tokenData.texture.src = tokenDocument.actor.img;
 		if (tokenDocument.actor.system.visions?.astral?.isActive) tokenData = await SR5_EntityHelpers.getAstralVisionData(tokenData);
@@ -225,6 +227,7 @@ export const registerHooks = function () {
 	});
 
 	Hooks.on("updateToken", async function(tokenDocument, change) {
+		//if (game.combat) SR5Combat.changeInitInCombatHelper(tokenDocument.id);
 		if (change.x || change.y) {
 			SR5_EffectArea.tokenAura(tokenDocument);
 			if (game.user.isGM){
@@ -268,29 +271,33 @@ export const registerHooks = function () {
 	});
 
 	Hooks.on("updateItem", async(document, data, options, userId) => {
-		if (document.isOwned && game.combat) SR5Combat.changeInitInCombat(document.actor);
+		if (document.isOwned && game.combat && game.user?.isGM) SR5Combat.changeInitInCombatHelper(document.actor.id);
 		
 		//Keep agent condition monitor synchro with owner deck
 		if(document.type === "itemDevice" && data.system.conditionMonitors?.matrix){
 			if (document.parent?.type === "actorPc" || document.parent?.type === "actorGrunt"){
 				for (let a of game.actors) {
-					if(a.type === "actorAgent" && a.system.creatorId === document.parent.id) await SR5Actor.keepAgentMonitorSynchro(a);
+					if(a.type === "actorAgent" && a.system.creatorId === document.parent.id) await SR5_ActorHelper.keepAgentMonitorSynchro(a);
 				}
 			} 
 		}
 	});
 
 	Hooks.on("updateActor", async(document, data, options, userId) => {
-		if (game.combat) SR5Combat.changeInitInCombat(document);
+		if (game.combat && game.user?.isGM) {
+			let actorId = document.id;
+			if (document.isToken) actorId = document.token.id;
+			SR5Combat.changeInitInCombatHelper(actorId);
+		}
 
 		//Keep deck condition monitor synchro with agent condition monitor
 		if (document.type === "actorAgent" && data.system.conditionMonitors?.matrix){
-			await SR5Actor.keepDeckSynchroWithAgent(document);
+			await SR5_ActorHelper.keepDeckSynchroWithAgent(document);
 		}
 
 		//Keep edge monitor synchro with tokens
 		if (document.type === "actorGrunt" && data.system?.conditionMonitors?.edge){
-			await SR5Actor.keepEdgeSynchroWithGrunt(document);
+			await SR5_ActorHelper.keepEdgeSynchroWithGrunt(document);
 		}
 		//let truc = document.effects.find(e => e.origin = "linkLock")
 		//if (truc) await document.deleteEmbeddedDocuments('ActiveEffect', [truc.id]);
@@ -304,8 +311,8 @@ export const registerHooks = function () {
 			}
 			if (item.type === "itemEffect"){
 				if (item.system.hasEffectOnItem && item.parent){
-					if (item.parent.isToken) await SR5Actor.deleteItemEffectFromItem(item.parent.token.id, item.system.ownerItem);
-					else await SR5Actor.deleteItemEffectFromItem(item.parent.id, item.system.ownerItem);
+					if (item.parent.isToken) await SR5_ActorHelper.deleteItemEffectFromItem(item.parent.token.id, item.system.ownerItem);
+					else await SR5_ActorHelper.deleteItemEffectFromItem(item.parent.id, item.system.ownerItem);
 				}
 			}
 
@@ -322,16 +329,16 @@ export const registerHooks = function () {
 		}
 		//Remove item from PAN if it was slaved
 		if (item.system.isSlavedToPan){
-			SR5Actor.deleteItemFromPan(item.uuid, item.system.panMaster, null);
+			SR5_ActorHelper.deleteItemFromPan(item.uuid, item.system.panMaster, null);
 		}
 	});
 
 	Hooks.on("deleteActiveEffect", async (effect) =>{
-		if (effect.flags.core?.statusId === "prone") {
+		if (effect.flags.core?.statusId === "prone" && (effect.testUserPermission(game.user, 3) || (game.user?.isGM))) {
 			let itemEffect = effect.parent.items.find(i => i.type === "itemEffect" && i.system.type === "prone");
 			if (itemEffect) {
-				if (itemEffect.parent.isToken) await SR5Actor.deleteItemEffectLinkedToActiveEffect(itemEffect.parent.token.id, itemEffect.id);
-				else await SR5Actor.deleteItemEffectLinkedToActiveEffect(itemEffect.parent.id, itemEffect.id);
+				if (itemEffect.parent.isToken) await SR5_ActorHelper.deleteItemEffectLinkedToActiveEffect(itemEffect.parent.token.id, itemEffect.id);
+				else await SR5_ActorHelper.deleteItemEffectLinkedToActiveEffect(itemEffect.parent.id, itemEffect.id);
 			}
 		}
 	});
