+++ scripts/rolls/roll-test-helper.js
@@ -58,7 +58,7 @@ export class SR5_RollTestHelper {
 
     //Remove 1 edge from actor
 	static async removeEdgeFromActor(messageData, actor) {
-		if (messageData.actorType === "actorSpirit") {
+		if (actor.type === "actorSpirit") {
 			let creator = SR5_EntityHelpers.getRealActorFromID(actor.system.creatorId);
 			creator.update({ "system.conditionMonitors.edge.actual.base": creator.system.conditionMonitors.edge.actual.base + 1 });
 		} else {
