+++ templates/rolls/roll-card.html
@@ -67,9 +67,7 @@
 		{{#if (ne test.type 'falseTest')}}
 		<div class="SR-CardButtonHit grid">
 		<div class="col-2">
-			{{#if chatCard.canEditResult}}
 				<a class="chat-button-gm edit-toggle" title="{{localize 'SR5.EditResult'}}"><i class="fas fa-edit"></i></a>
-			{{/if}}
 			{{#if chatCard.templatePlace}}
 				<a class="messageAction owner" title="{{localize 'SR5.PlaceTemplate'}}" data-action="nonOpposedTest" data-type="templatePlace"><i class="fas fa-bullseye"></i></a>
 			{{/if}}
